# go_api/docs/swagger.yaml
definitions:
  dto.AssignPermissionRequest:
    properties:
      permission_id:
        type: integer
    required:
    - permission_id
    type: object
...
  dto.CreatePermissionRequest:
    properties:
      code:
        type: string
      description:
        type: string
      group:
        type: string
      name:
        type: string
    required:
    - code
    - name
    type: object
  dto.CreateRoleRequest:
    properties:
      role_name:
        type: string
    required:
    - role_name
    type: object
...
  models.PaginatedPermissionsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
    type: object
  models.PaginatedRolesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.RoleWithCountResponse'
        type: array
    type: object
  models.PaginatedRoomsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.Room'
        type: array
    type: object
  models.PaginatedStudyFieldsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.StudyFieldListSummaryResponse'
        type: array
    type: object
  models.PaginatedStudyProgramsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.StudyProgramDetailResponse'
        type: array
    type: object
...
  models.Permission:
    properties:
      code:
        type: string
      description:
        type: string
      group:
        type: string
      id:
        type: integer
      name:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
    type: object
  models.Role:
    properties:
      id:
        type: integer
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      role_name:
        type: string
    type: object
  models.RoleWithCountResponse:
    properties:
      id:
        type: integer
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      role_name:
        type: string
      users_count:
        type: integer
    type: object
  models.Room:
    properties:
      building_id:
        type: integer
      faculty_id:
        type: integer
      id:
        type: integer
      pc_amount:
        type: integer
      projector_availability:
        type: boolean
      room_capacity:
        type: integer
      room_name:
        type: string
      unit_id:
        type: integer
    type: object
  models.SemesterSummary:
    properties:
      courses_count:
        type: integer
      ects_sum:
        type: integer
      semester_number:
        type: integer
    type: object
  models.StudyField:
    properties:
      degree:
        type: string
      faculty_id:
        type: integer
      field_name:
        type: string
      id:
        type: integer
      language:
        type: string
      mode:
        type: string
    type: object
  models.StudyFieldListSummaryResponse:
    properties:
      degree:
        type: string
      elective_blocks_count:
        type: integer
      faculty:
        type: integer
      field_name:
        type: string
      id:
        type: integer
      language:
        type: string
      mode:
        type: string
      programs_count:
        type: integer
      semesters_count:
        type: integer
      specializations_count:
        type: integer
    type: object
  models.StudyProgram:
    properties:
      id:
        type: integer
      program_name:
        type: string
      start_year:
        type: string
      study_field:
        type: integer
    type: object
  models.StudyProgramDetailResponse:
    properties:
      id:
        type: integer
      program_name:
        type: string
      semester_summary:
        items:
          $ref: '#/definitions/models.SemesterSummary'
        type: array
      semesters_count:
        type: integer
      start_year:
        type: string
      study_field:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/permissions:
    get:
      description: Returns list of permissions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedPermissionsResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get permissions
      tags:
      - permissions
    post:
      consumes:
      - application/json
      description: Creates a new permission
      parameters:
      - description: Permission data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Permission'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create permission
      tags:
      - permissions
  /api/v1/roles:
    get:
      description: Returns list of roles with permissions and user count
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedRolesResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Creates a new role
      parameters:
      - description: Role data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create role
      tags:
      - roles
  /api/v1/roles/{id}/permissions:
    post:
      consumes:
      - application/json
      description: Assigns a permission to a role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Permission assignment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AssignPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Assign permission to role
      tags:
      - roles
  /api/v1/rooms:
    get:
      description: Returns list of rooms
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedRoomsResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get rooms
      tags:
      - rooms
    post:
      consumes:
      - application/json
      description: Creates a new room. ID is generated by database and cannot be provided
        by client.
      parameters:
      - description: Room data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Room'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              $ref: '#/definitions/models.Room'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create room
      tags:
      - rooms
  /api/v1/study-fields:
    get:
      description: Returns study fields with aggregated counts (majors, programs,
        elective blocks, semesters)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedStudyFieldsResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get study fields
      tags:
      - study-fields
    post:
      consumes:
      - application/json
      description: Creates a new study field.
      parameters:
      - description: Study field data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.StudyField'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.StudyField'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create study field
      tags:
      - study-fields
  /api/v1/study-programs:
    get:
      description: Returns study programs with semester statistics (courses count
        and ECTS sum per semester)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedStudyProgramsResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get study programs
      tags:
      - study-programs
    post:
      consumes:
      - application/json
      description: Creates a new study program. ID is generated by database and cannot
        be provided by client.
      parameters:
      - description: Study program data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.StudyProgram'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.StudyProgram'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create study program
      tags:
      - study-programs
...
swagger: "2.0"