# go_api/internal/handlers/v1/room_handler.go
package handlers

import (
	"net/http"

	"github.com/gin-gonic/gin"

	"go_api/db"
	"go_api/internal/models"
)

// GetRooms godoc
// @Summary Get rooms
// @Description Returns list of rooms
// @Tags rooms
// @Produce json
// @Success 200 {object} models.PaginatedRoomsResponse
// @Failure 500 {object} map[string]string
// @Router /api/v1/rooms [get]
func GetRooms(c *gin.Context) {
	var rooms []models.Room

	if err := db.DB.Order("id").Find(&rooms).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, models.PaginatedRoomsResponse{
		Items: rooms,
	})
}

// CreateRoom godoc
// @Summary Create room
// @Description Creates a new room. ID is generated by database and cannot be provided by client.
// @Tags rooms
// @Accept json
// @Produce json
// @Param request body models.Room true "Room data"
// @Success 201 {object} map[string]models.Room
// @Failure 400 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /api/v1/rooms [post]
func CreateRoom(c *gin.Context) {
	var room models.Room

	if err := c.ShouldBindJSON(&room); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": err.Error(),
		})
		return
	}

	if err := db.DB.Create(&room).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{
			"error": err.Error(),
		})
		return
	}

	c.JSON(http.StatusCreated, gin.H{
		"data": room,
	})
}